requirements.txt
flask
torch
torchvision
Pillow
matplotlib
tqdm
